Refactor toolbox structure to prepare for Federal Senate dataset
As we were doing with Rosie, we are now preparing the toolbox to receive a new dataset.
What need to be done:
- [] Add the script to get the Federal Senate datasets.
- [] Add the datasets to Amazon S3
