+++ tests/test_reimbursements.py
@@ -1,5 +1,5 @@
 from unittest import TestCase, main
-from serenata_toolbox.reimbursements import Reimbursements
+from serenata_toolbox.chamber_of_deputies.reimbursements import Reimbursements
 import numpy as np
 from numpy.testing import assert_array_equal
 import pandas as pd
